Move request to background task
